ENGR00132615-1 esdhc: auto switch clock to 20Mhz when AHB is 24Mhz
The issue we found in 35 kernel is that esdhc will easily get failed
with crc error when ahb bus clock is down to 24Mhz in low power mode.

There's still no root cause found yet.

This is a workaround that automatically switching the esdhc clock
to 20Mhz when AHB is 24 Mhz(LP mode) to prevent this issue to happen
and switching back to the max_clk of esdhc when exiting LP mode.

Signed-off-by: Aisheng.Dong <b29396@freescale.com>