ENGR00132615-2 mx50: delete AHB_HIGH_SET_POINT flag of esdhc clock
This is another work-around for esdhc crc error issue when AHB is
down to 24Mhz in low power mode.
But an expensive one since AHB will go back to 133Mhz and it will switch
the entire LP domain frequencies whenever the SDHC is accessed

With the workaround of switching esdhc clock to 20Mhz when AHB is 24Mhz,
we do not need to have this flag set anymore.

Signed-off-by: Aisheng.Dong <b29396@freescale.com>