ENGR00142681 dvfs core: set high cpu freq when GPU is running on Android
GPU benchmark performance drop after DVFS core enabled by default on Android,
and we found it is because of low CPU frequency when doing benchmark.
Under most of the 3D applications cases, high cpu freq is needed to get better
performance, so we just disable cpu freq adjustment when GPU running
to fix this issue.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>