ENGR00151685 Android: Fix dvfs have too much irq in some case
After apply this patch:
dvfs core: set high cpu freq when GPU is running on Android.

we observe events/0 kernel thread will comsuming about 50%
CPU usage, also IRQ will comsume 15% CPU usage.

The root cause is after apply the GPU patch, we don't let dvfs
adjust the cpu frequency when GPU is running,
so the dvfs will keep coming irq let us adjust the frequency,
but we still ignore it, and the irq keep commming.

So add code to fix it by:

In this case, we disable DVFS's irq and poll the DVFS's
status every 100ms(may should adjust), if the gpu's clock
is not being used, it will exit the loop mode and
enable the dvfs's irq again.

Signed-off-by: Zhang Jiejing <jiejing.zhang@freescale.com>